+++ lib/write12.c
@@ -8,10 +8,10 @@
 #define REGPARM(x)
 #endif
 
-int REGPARM(1) __write1 (const char* s) {
+ssize_t REGPARM(1) __write1 (const char* s) {
   return write(1, s, strlen(s));
 }
 
-int REGPARM(1) __write2 (const char* s) {
+ssize_t REGPARM(1) __write2 (const char* s) {
   return write(2, s, strlen(s));
 }
