+++ libpthread/pthread_sys_read.c
@@ -3,7 +3,7 @@
 #include <pthread.h>
 #include "thread_internal.h"
 
-int read(int fd, void *buf, size_t count) {
+ssize_t read(int fd, void *buf, size_t count) {
   __TEST_CANCEL();
   return __libc_read(fd,buf,count);
 }
